refactor: migrate internal callers to molecule APIs
Ported-from: f0d7ad65b77a1834a717237dd87c2d5679713d2e
