+++ scripts/sascore_bench/run_benchmark.py
@@ -62,7 +62,7 @@ from typing import Any
 import yaml
 from tqdm.auto import tqdm
 
-from synplan.chem.utils import mol_from_smiles
+from synplan.chem.molecule.standardization import mol_from_smiles
 from synplan.mcts.search import _iter_target_smiles
 from synplan.mcts.tree import Tree
 from synplan.utils.config import RolloutEvaluationConfig, TreeConfig
