+++ synplan/mcts/evaluation.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 
 import torch
 
-from synplan.chem.precursor import Precursor, compose_precursors
+from synplan.chem.molecule.precursor import Precursor, compose_precursors
 from synplan.chem.rdkit_utils import RDKitScore
 from synplan.chem.reaction.rules import POLICY_SOURCE_NAME
 from synplan.ml.networks.checkpoint import load_network_from_checkpoint
