+++ synplan/mcts/policy/composite.py
@@ -10,7 +10,7 @@ import torch
 from synplan.mcts.policy.base import Policy
 
 if TYPE_CHECKING:
-    from synplan.chem.precursor import Precursor
+    from synplan.chem.molecule.precursor import Precursor
     from synplan.chem.reaction import CanonicalRetroReactor
