+++ synplan/mcts/policy/template_based.py
@@ -22,7 +22,7 @@ from synplan.ml.featurization.rules import query_cgr_graphs_from_smarts
 if TYPE_CHECKING:
     import torch_geometric.data
 
-    from synplan.chem.precursor import Precursor
+    from synplan.chem.molecule.precursor import Precursor
     from synplan.chem.reaction import CanonicalRetroReactor
     from synplan.ml.networks.policy.linear import (
         FilteringPolicyNetwork,
