+++ synplan/utils/loading.py
@@ -18,12 +18,11 @@ from chython.reactor.reactor import Reactor
 from huggingface_hub import hf_hub_download, snapshot_download
 from tqdm.auto import tqdm
 
+from synplan.chem.molecule.io import standardize_sdf_text, standardize_smiles_batch
 from synplan.chem.reaction import CanonicalRetroReactor
 from synplan.chem.utils import (
     AtomMappingCheck,
     reaction_string_mapping_status,
-    standardize_sdf_text,
-    standardize_smiles_batch,
 )
 from synplan.ml.networks.value import ValueNetwork
 from synplan.utils.config import ReactorConfig
